Configure Poetry for Devcontainer (#2)
# Why Are We Doing This

When I tried to use the devcontainer configuration, as it was, as a
starting point in a different project it didn't work as expected. I was
getting errors when I ran poetry commands. These changes help keep the
external poetry separate from the one inside the container.

## Always Don't Do What Barry Did

Even though it is in the container, we still want to use a virtualenv
for the application and separate environment for `poetry`.
When Barry insisted on doing what he did, setting `virtualenvs.create =
false`, we ran into the issue described in [this
issue][poetry-isolation] and then we did the work around. That's a lot
of non-standard effort to avoid typing `poetry run` inside the
container.

In this image, `poetry` is installed via `pipx`. There are several other
tools installed in a similar manner that are already a part of the base
image. Those tools should probably be pinned. They aren't yet; well,
poetry isn't. IDK about the tools that come with the image.

## All Your Python Are Belong To Us

I dropped the universal image (i.e. "All the things!") that I was
initially using and picked one focused on Python. Still a Codespaces /
Devcontainer image. It comes with 3.11 and then I installed 3.8, 3.9,
3.10, and 3.12. They build in parallel in a multi-stage build, so it is
as fast as it can be. If that is too much there are still things that
can be done to reduce the build time - like use the `COPY --link`
command or publish the image or use volume mounts for better disk
performance etc.

[poetry-isolation]:
https://github.com/python-poetry/poetry/issues/6397#issuecomment-1236327500